added context discord refactor
